add: update and delete alien

+++ src/main/java/com/divergenny/springbootrest/AlienRepository.java
@@ -2,7 +2,7 @@ package com.divergenny.springbootrest;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AlienRepository extends CrudRepository<Alien, Integer> {
+public interface AlienRepository extends CrudRepository<Alien, Long> {
     Alien findAlienById(Long id);
 
     Alien save(Alien alien);
